Fill out preliminary GitHub data download methods
The gateway.go file holds the "getter" methods for issues and PRs found
on GitHub. Currently, the "Event"-based getters are not final as their
need has yet to be confirmed. Additionally, this commit also holds
adjustments that were made to the directory name as hyphens are not
acceptable in Go package names.