Backtest setup intial checkin
Gateway and conflation is now integrated into the backtest with support for saving to disk.
Note: Above code will likely be cleaned up as this is an initial checkin
Note: extractIssueId method in OneToOne conflation needed to be changed as IssueURL field is not actually the issue url.
Note: A github service account may need to be created for API access