Temporary code additions for model examination
This commit includes a great deal of TEMPORARY code that will be removed
once the final product is ready. Most of the additions involve printint out
values so that potential errors can be evaluated. Most of the changes present
in this commit will be undone.

Note: the primary issue being investigated here is looking into log scores
and how to best to handle them.
WARNING: THIS COMMIT CONTAINS TEMPORARY CODE