Restructure cross validation functionality
Description:
This is a comprehensive rebuild of folding operations. Most of the code here
involves unifying operations and changing how they generate output so that it
is more suited to a logging-type use (as opposed to simple commandline output).
This is an extensive rebuild of the fold functions; all three fold functions
now have uniformed input/output generation and have been refactored heavily.

Warnings:
A number of general changes have been made to test files and other dependent
source files.
NOTE: several test files have been adjusted for the new functionality; they
have NOT been comprehensively review for soundness.
NOTE: this code built and passed the 'go test' for the repository.
NOTE: all of the affected files have had 'gofmt' run on them.
NOTE: the actual calls to the various folds have yet to be implemented.
NOTE: backtest runs have not been made with this commit code.

Build well