Build initial skeleton for Scenario3A
Description:
This is a first attempt at a possible scenario structure for use in the
variadic filtering functionality. Two unit tests are included with the initial
check-in.

Warnings:
NOTE: the tests are not comprehensive; one is still needed to ensure it handles
      being passed a non-pull request object.
NOTE: there is a great deal of in-code comments both for TODO and temporary
      code removal purposes.

Build well