Construct initial scenario1b
Description:
This is a placeholder commit in order to update the branch with code from
master. Included is an outline for scenario1b as well as documentation updates
in a variety of files.

Warnings:
Note: the unit test for the new scenario1b is currently failing intentionally.
Note: several files have been removed (log outputs); they should not be
      included in commits going forward.

Build well