Implement all helper functions in model
Description:
This contains a wide variety of repairs / additions. The word stemmer has been implemented in the core model methods. Additionally, changes have been made to unit tests and the associated tossing and folding functionality.

Warnings:
Note: the ultimate model output has not been reviewed for accuracy.
Note: all unit tests are passing but have not been reviewed for completeness.
Note: there is missing in-code documentation.
Note: there are notes for future refactoring (with regards to the logger and model assets).

Build well