Implement Scenario4a filter code
Description:
This is another addition to the collection of interchangeable scenario filters.
Scneario4a looks to identify issues with a specific number of words in their
text bodies.

Note: the number of words is specified in the instantiation of the Scenario4a
struct value.
Note: all tests are passing.

Build well