Restructure logging output
Description:
This implements 1) the new logging output subdirectories and 2) a unit test for
logging. There is included logic for handling logic for flags indicating
backtests for CoralReefLogger struct.

Note: the unit test is passing.
Note: output for caches/ has not been tests (although in theory the new output
      directory string should generate the files in the desired location).

Build well