Initial broad refactoring of project assets
Description:
This was the initial restructuring of the overall project; the
primary focus was on removing assets from the Bhattacharya model
and "globalizing" them for repeat / shared usage between future
models.
Additionally, this was also the starting point for rebuilding
individual function / method parameters with a more "global"
project mindset.

Note:
This was a squashed commit that encompasses a large number of
individual file changes.

Build well