Change high-level parameters to ExpandedIssue
Description:
The highest-level Model methods have had their parameters
changed to the global asset ExpandedIssue for modularity;
this is to help focus on ExpandedIssue as the primary asset
that is passed initially into every model.
Going forward, additional fields can be added to the
ExpandedIssue struct should the need arise without impacting
existing architecture.

Note:
The unit tests has begun to be updated appropriately to
reflect this change.

Build well