Attempt callback handler unit test build
Description:
The broader callback handler has been removed as well as the respective unit
test and placed in separate files.
Additionally, the HTML will no present "Heupr" in the webpage tab title.
Unnecessary assets / tests have been removed as the handler is now depreciated.

Due to time considerations, and the potential "plug-n-play" nature of the
GitHub client in the broader program, we are temporarily removing these unit
tests to focus on other aspects of the project.

Note: These unit tests need to be implemented in the future.

Build well