Add assets for collecting pull requests
Description:
Since the models will likely be training on PRs as well as issues, this helper
method has been added to the HeuprServer; this was a planned addition.

Note: Some code has been commented out (likely going to be replaced)
Note: Unit tests still need to be built for code in this commit.
Note: Existing unit tests appear to be passing.

Build well