Move user signup assets to separate directory
Description:
Separating the data flow and user initialization is part of the pivot we've
taken in our project structure. This is important for the simplification of the
data collection/manipulation process as well as program security.

Note: Unit tests are passing within the onboarder/ directory
Note: Tests fail in the frontend/ directory (due to variables missing)
Note: All code is properly formatted
Note: More code needs to be changed for this swap to be complete
Note: Possibly transitioning to the gob package shortly

Build well