Implement database structure assets
Description:
This is a broad rebuilding of the database operations. The core is the MemSQL
relational database and there are several "coupled" components that we have
built into it. These include:
- front-end population
- optimized insert pooling
- back-end reading
- worker/dispatcher training logic
However, it should be noted that this code is currently incomplete and there
is a lot of "in progress" content.

Note: Some unit tests are failing at this point.
Note: Code coverage is incomplete.

Build well