Introduce MemSQL and refactor pipeline
Description:
This is a broad commit that collects several different aspects (primarily of the backend
logic) into a single update.

MemSQL is now the core of the data pipeline and insert/fetch logic has been updated
accordingly. This pipeline is likely subject to significant change as the channel
logic is still under evaluation.

The pipeline has been split into distinct frontend and backend sections with
dedicated logic and independent operations. This is to help maintain ease of updating
either end without requiring both to be operational concurrently.

Note: This code is generally, but not completely, passing unit tests.
Note: Most code here will need factoring/cleaning prior to beta deployment.