Build activation service helper methods
Description:
A check for database record continuity has been implemented as a method (split
into the continuity.go file); a test file has been included but is currently
empty.

Because there may be interruptions to the ingestor server, a Restart() method
was needed to update the database with information that may have been missed
while the server was offline. Because models will likely eventually account for
more than just issues/pulls, this method is subject to change. Personally,
I'm going to revisit these methods (Restart and ContinuityCheck) to at the very
least heavily refactor out some of the redundant functionality found within the
method.

Also, a somewhat misleading (at least to mean) variable name has been adjusted
for the backend query logic. This is not impactful to the code functionality,
only the readability.

Note: NONE OF THIS CODE IS UNIT TEST COVERED.
Note: Code is only passing build checks (not logic / completeness).