Add initial website content
Description:
This checkin includes an initial version of the Hugo-generated static website
in the website/ subdirectory (this is subject to change / amendment). This was
done to test the signup workflow which was partially working (there were a few
workarounds that were required due to the project not yet being registered with
GitHub).

Additionally, several logging and error handling structures have been tweaked
for a more logical flow.

Note: Code is passing (mainHandler unit test is commented out).
Note: Full signup run-throughs have yet to be completed.
Note: Some aesthetics (e.g. repo selection page) need to be written.