Refactor Timer method caller
Description:
Timer needed to be cleaned up a bit and hopefully this helps. There is now a
select statement to open up the potential for a graceful shutdown to be built
in the future; also the Dispatcher and Collector have been moved to different
scopes - I think this makes sense because otherwise there would be Workers
spun up on every tick.

Note: Code is passing the unit tests (not tests on Timer though).
Note: This directory needs more unit tests build.
Note: Fixes #24.