Refactor ingestor logic
Description:
There are several areas of the logic within the
ingestor server (and related methods/functions)
that need to be improved for both code
clarity/efficiency and unit testability.

Unit tests have not yet been added as this is
just a check-in to merge into the master branch.

Several of the internal URLs that the various
handlers would respond to were inconsistent and
in some cases incorrect. I made them more
uniform and also made some adjustments to some
erroneous FormValue parsing logic as it did not
conform to the correct structure that the
frontend was putting it into.

Note: Incomplete unit test coverage.
Note: Code may not be completely stable.
Note: Progress towards #25.