Issue Double Assigment Fix
Description: Modified the query in the backend to pull down the latest revision of a github event. This will prevent redundant assignments during a recovery scenario. Also, I removed the DB open/close from the restart and continuity as this given DB object is shared by other go-routines and it cannot be closed without causing other go-routines to fail.