Rewrite SQL continuity query
Description:
The previous query was not going to work since it needed to account for
multiple repo IDs existing in the database (the PARTITION handles this); also,
the "looping logic" has been updated to handle how results are returned (since
it is now returned with the start/end numbers NOT included in the missing
numbers).

Note: The query was tested independently of this code.
Note: The unit test has been updated.
Note: This test does not account for multiple repo IDs.