Build automatic whitelist updater
Description:
The function here will ultimately be implemented within the
githubCallbackHandler method and act as a "gateway" into the platform. This is
pretty bare-bones and some of the stuff that we are currently storing in the
BoltDB instance may change (but since we basically just want to count entries
and potentially see the names of the repos quickly, it may be enough).

Note: Unit testing non-comprehensive.
Note: The manual updater needs to be built still.
Note: Currently, the existing unit test is passing.
Note: The AutomaticWhitelist has NOT been implemented in the production code.