Complete whitelist logic
Description:
The whitelist.toml file is now being referenced correctly as if it were a "VIP"
list; a check is in place to see that the requesting individual/repo is in the
TOML file.

Note: A unit test has been added for the boolean check helper function.
Note: The unit tests are not comprehensive.
Note: Code has been formatted correctly.