Refine limit selection logic
Description:
This includes adjustments that were made to the logic (from a prior commit) as
well as updates to the BoltDB storage functions. There is now capacity to store
the limit ("cutoff") date in the file for use in recovery scenarios.

Note: Contributes to #33.
Note: All unit tests are passing in frontend/ and backend/.
Note: Code has been formatted properly.
Note: End-to-end tests have not been run on this code yet.