Update frontend error messaging
Description:
Most of the errors now post into Slack as opposed to returning to the client.

Also note that this includes moving the BotlDB operations into a separate
goroutine in order to avoid starving the resources. Unit tests passed and a
test on localhost failed (as expected) and appropriately posted into Slack. I
think the goroutine is okay as well but I'm not 100% sure.

Note: Code formatted properly.
Note: Code is passing available unit tests.