Adding Prod Build Flag to Frontend
Description: When building the frontend with the commands "go build" a newly added PROD const will be set to false at compile time. This const is used as a flag to determine which server IP to utilize and which ssl key to utilize. This PROD const is unique to the Frontend.