Add frontend server graceful shutdown
Description:
This may be overkill for the frontend following the switch to GitHub
Integrations. Additionally, a field has been added to the FrontendServer struct
to allow for the Shutdown method to be called correctly.

This commit also contains a refactoring of the prod/dev loop when starting up
servers; lastly, the outlines of a unit test for LaunchServer is included (it
may be better to just unit test Start/Stop given the need to build .pem files).

Note: Code is formatted appropriately
Note: Code appears to be working correctly
Note: No unit tests have been implemented for the server.go file