Fix initial unit test outline
Description:
The 3rd party module was generating a print output which needed to be
suppressed. Also (from the prior commit), the BoltDB assets have been commented
out due to the impending depreciation.

Note: Unit test is incomplete but not breaking
Note: Temporary .pem files are not being removed (known issue)