Replacing OAuth with Github Apps Integration. (Initial Checkin)
Description: 
Given the webhook & Github specific nature of the apps integrations authentication workflow, the repo signup logic now lives in the ingestor triggered via an InstallationEvent. This change in the ingestor rendered BoltDB, the entire AuthenticationServer, and the frontend signup logic obsolete. The Frontend is now just a dummy facade. The hooker, whitelist, storage/boltdb, and authentication logic have been deleted from the Frontend. The AuthenticationServer is deleted. As a result the ingestor is the gatekeeper to a repo signup. This new responsibility for the ingestor means the ingestor now sends out a repo signup message directly to the BackendServer. The Ingestor now assumes the role of the now deleted/obsolete AuthenticationServer. With BoltDB gone the Backend, and Ingestor now read exclusively from MemSQL. This lead to tweaks in the data access layer of the Backend's and Ingestor's recovery/ restart logic. 

Notes:
Note: This logic does not support integration deletion. (This is still WIP and is needed before we can release this)
Note: Most but not all unit tests have been updated accordingly.
Note: All obsolete unit tests have been deleted. 
Note: Performance optimizations have been made to the ingestor. No longer creating redundant MemSql DB struct instances.
Note: This first domino that fell is in pipeline/ingestor/worker.go.
Note: The Replay Server was tweaked to correct build errors.