OAuth to Integrations (Checkin 3) Checkpoint
Description: Added "Hello, I'm Heupr"  Issue/ welcome page. This includes both the content and the settings parsing logic. As this is a checkpoint checkin what is still outstanding is what's more important to focus on. 
TODOs:
TODO: When a user without repo contributor status attempts to edit a Heupr setting. Generate an error messaging in the reply and CC an admin level contributor. Now, we may also want to throttle our error messages in this case as a malicious user could use this mechanism against us and spam the repo.  
TODO: When the user gives the green light "yes" to the settings insert the settings  into the database. 
TODO: Read Heupr settings into the backend. 
TODO: When a parsing error occurs display the settings to the user and let them know an error occured. Low priority as parsing tends to silently fail anyway.