Build unit test for repo initializer check logic
Description:
The unit test basically provides a check against the three possible scenarios
that the method evaluates to. There is probably a better way, but I just
included the necessary stub in the unit test for the AddRepo DataAccess
interface (overwriting one of the methods).

I also made some slight naming changes to the repo_initializer.go methods.

Note: Not all golint standards are met in this check-in
Note: Unit tests are passing (but not all code is covered)