Building unit test for backend activator
Description:
This isn't my best/favorite unit test and will likely need some tweaks to bring
it up to standard. There was a need to "catch" the logger output and evaluate
them (given the structure of the method).

Note: Unit tests are passing
Note: This unit test needs to be checked/evaluated to make sure it is
      accomplishing what it's supposed to