Unit test for backend collector function
Description:
There was an empty test file for the collector function so I threw in a quick
test. It's not particularly comprehensive (not that it really needs to be given
the function's simplicity), and only evaluates a correct scenario.

Additionally, some golint requirements have been updated specifically as they
relate to the collector.go file.

Note: Unit test coverage expansion
Note: Unit tests in backend are passing