Adding new website files in temp storage
Description:
These files have gotten to the point that I am uncomfortable with them NOT
being in GitHub so I've tossed them into a temporary folder (just in case we
need to recompile the Hugo site for some reason before we deploy these new
pages). These are a work in progress so the HTML files themselves still need
work as well.

Note: WIP check-in