Progress update on website console
Description:
There are still sections that need to be retooled and some functionality is
not working. For example, the cloned dropdown selections do not actually
respond to user input (at least as far as I can tell) for some reason and a
"save" function still needs to be built.

Note: Page/functionality is incomplete
Note: WIP check-in