Update reference for cloning
Description:
Previously, the clone in the addBucket function was pointing at an ID but
because the entire bucket was being cloned, it was also copying the ID (which
is against best practice). This change now has it referencing just the first
result that is found matching the given class.

Note: Refactoring
Note: Functionality is still missing in this check-in