Add check for unique bucket names
Description:
This is (likely) a necessary feature for rest of the pipeline to more easily
process buckets. However, this implementation is pretty messy and can
DEFINITELY be fixed but I'm fed up with jQuery/JS for now.

Also, there's a minor update to the landing page (more to come).

Note: Progress on #48 (I believe only a minor change is needed to the "action"
      attribute in order to have the "submit" function working as needed)
Note: This does NOT check for empty buckets (which is also a check that needs
      to be implemented soon)