Rebuilding frontend server logic
Description:
A new series of handlers needs to be built for the website and this is a step
towards that task.

Several of the site pages are static so there is now a "generator" which
produces a http.Handler for a given filepath - unit tests have also been
updated for this piece. I have NOT implemented the handlers on the server yet
(I'll do that in a subsequent check-in).

There is also a new website2/ directory (temporary name) which stores the HTML
files for clarity. I've added a few new files (some of which do not yet have
content) as well.

Note: Unit tests passing (coverage incomplete)
Note: WIP check-in
Note: Progress towards #51 and #49