Update console to allow dynamic repo selection
Description:
Since multiple repos could have Heupr installations for the given
authenticated user, there needs to be the ability to select which is
currently being adjusted.

The temporary JS serves as a visualization of what would be there but
the actual flow will be different (see #48 for details). For example,
there may need to be an additional <form> element in place to account
for repo selection.

Note: WIP check-in
Note: Progress on #48
Note: Some JS is still in temporary state