Expand GET request logic in console handler
Description:
Various pieces of logic related to handling GET, creating OAuth
clients, and validating requests are included. The logic is still
incomplete however and needs to be fleshed out (see #51 for details)
to be fully functional.

Also, the unit test is passing at this point but will need to be
gradually updated as well (and expanded for different scenarios).

Note: WIP check-in
Note: Progress on #51
Note: Production ClientID/Secret have been left out for now
Note: Unit tests are passing (but incomplete)
Note: This includes commented/depreciated code