Partial update to template arrangement
Description:
Templates will be used to render the webpages and the first part of
that transition is here. Some JavaScript will likely need to be
tested/updated to account for the selection options made in the
dropdown.

Note: WIP check-in
Note: Progress on #51
Note: This code is very incomplete and includes comments, notes, and
      TODOs for the feature