Expand GET processing logic
Description:
At this point, the GET method receiving logic should be relatively
complete (although the actual ListUserRepos method does still need
to be checked to make sure that it is the right one). The logic now
contains template support and everything appears to be compiling
correctly.

There are also some minor/initial updates to the docs webpage
content.

Note: WIP check-in
Note: Progress on #51
Note: Unit tests are passing (but incomplete)
Note: Some of the methods here may NOT be functionally complete