Add GET method storage functionality
Description:
Information collected on the GitHub pulldown for populating into
the webpage template needed to be made available to the POST
methods; to handle this, gob and output files are used to store it
for reference as needed. Some of this implementation (e.g. structs
created for storage) may need to be tweaked to be more idiomatic.

Also, the ListUserRepos has been confirmed to be the correct method
to use in pulling down the relevant repos.

The unit tests have been updated (and now also include the capacity
to remove the files created during the test) and now use negative
integers to differentiate from actual GitHub repos (see the dummy
handlers).

Note: WIP check-in
Note: Progress on #51
Note: Unit tests are passing
Note: Not all scenarios have been accounted for in the unit tests
      for the completed code
Note: Code for the POST methods has yet to be completed