Separate handler logic for two-step console flow
Description:
Since I'm switching over to a two-page workflow for the user to
access the console, the handlers needed to be updated (in this
case I'm simply splitting them into two sets of logic). There is
still a lot of logic that needs to be built/upgraded but this is
the gist of the new structure.

Note: WIP check-in
Note: Progress on #51
Note: Unit tests are passing (but not comprehensive)
Note: There are TODOs/NOTEs included for next tasks to handle