Add file handling logic to both handlers
Description:
For the repos handler there are now checks to make sure that the
file exists and the console handler now has some of the initial
logic required for pulling out GitHub/settings data.

A bunch of this logic (as it relates to the file management) will
change in the near future and likely become single-file per repo
with more logic performed after reading into memory.

Note: WIP check-in
Note: This is a VERY incomplete check-in
Note: Unit tests are passing but not comprehensive