Implement single file storage per signed up repo
Description:
Instead of creating two files per repo (one for GitHub data and one
for user settings data), I've adjusted the storage logic to create
or update a single file in-memory as needed. The unit tests have
been adjusted accordingly but I'm not 100% certain that they're
super comprehensive. Some of the handling of the gob logic may also
need to be tweaked (but it seems to be working).

Note: WIP check-in
Note: Progress on #51
Note: Unit tests are passing but not comprehensive
Note: Includes in-code NOTEs/TODOs