Change from Redirect to Error in responses
Description:
It appears to be more idiomatic to use the http.Error for statuses
not in the 2xx range so I've made that switch. Some of the error
messages are still being rendered to the client so that might need
to change.

Note: WIP check-in
Note: Includes incomplete code
Note: Unit tests are passing but are not comprehensive