Initial content for in-website docs page
Description:
This is an option if we choose to have all of our user resources
in the website and not in a separate GitHub repo/gist. The content
does not have all of the proper formatting that will be applied
later.

All of this content is subject to change.

Note: WIP check-in
Note: This is an incomplete check-in
Note: Progress on #50