Initial dynamic template label dropdown work
Description:
Rendering the console page required that the HTML file be updated
to support html/template; however, this is NOT complete as I have
yet to choose whether to 1) save an entire HTML file snapshot or
2) find out how to generate each of the dynamic label options
(from the prior choices of the user).

I ASSUME there's a way to do the latter (which would also require
placing the prior selections onto the dropdowns) and I would guess
that would be preferrable.

Note: WIP check-in
Note: Progress on #51
Note: This code (in the HTML and the handler) is incomplete
Note: Unit tests are passing but not comprehensive
Note: HTML file renders correctly in Firefox