Unifying Heupr style into single CSS file
Description:
All of the styling for the new website was embedded in the HTML and
I wanted to quickly move it out into a single, comprehensive CSS
file for ease of editing and uniformity. This does not change any
functionality on the website and was done purely for efficiency.

Note: All pages appear to be rendering correctly
Note: This includes incomplete work (in docs.html)