Add minor page content for uniformity
Description:
The <hr> has been added to most "simple" pages (it probably won't
be on the landing page). Also, the <div> element has been added
to replace the monotony of adding "content" tags to each <p>. There
will likely be a dedicated class for the docs items to provide
adequate spacing.

Note: WIP check-in
Note: Progress on #50