Broad website asset restructuring
Description:
The website was a little haphazardly organized and this is a move to
both clean it up and allow for more robust practices.

Some handlers have been broken into a separate file (and more will
likely be broken out in time into separate files). Also, the various
templates/static contents have been reorganized as well.

Templates are now used in all handlers (but these will likely need
to be adjusted as well); actual DRY practices still need to be added
(by implementing a base.html file and the appropriate handler logic)
but this is a step in that direction.

Note: WIP check-in
Note: Code is passing unit tests but not rendering CSS properly when
      run locally