Add direct file references for website assets
Description:
CSS and JS files are being added directly to the directory for
unity (I'm not sure if this is the most efficient solution but it
can be undone if need be). I've also included an initial base.html
file which will help adhere to DRY principles.

Note: WIP check-in
Note: This code has not been reviewed for functional correctness
      and is only serving as an outline