Update breaking website render
Description:
This is to fix the known/expected errors from a prior commit when
trying to render the site via a Go server. Below is a list of changes
that are included here.

- The file references in base.html have been updated correctly
- base.html still holds commented code (that will be depreciated)
- Both home.html and docs.html files have been updated to become
  defined templates (other HTML files still need this update)
- server.go now has an added static file handler
- A global baseHTML variable has been added for testing purposes
- Unit tests have been adjusted accordingly and are passing

Note: WIP check-in
Note: Progress on the frontend website
Note: Webpages render correctly from a demo server