Remove error check following template execution
Description:
Errors were thrown due to the error check returning an http.Error
that wrote twice to the ResponseWriter (which Go will not allow).
I'm not sure if removing the error check entirely is smart, but
I've done it to at least allow the errors to fire correctly.

Note: WIP check-in
Note: Progress on the frontend website