Remove error check after template execute
Description:
I did this in the other handlers previously (missed this one) due
to the potential resulting write to ResponseWriter in the Error
function being duplicate (which Go does not allow). This may not be
the appropriate answer but it seems to fit logically.

Note: WIP check-in
Note: Progress on the frontend website
Note: Removing unnecessary code
Note: A known unit test is failing currently but it seems like it is
      something wrong in the unit test itself