Additional frontend repairs and tweaks
Description:
Default labels were being improperly referenced in the HTML template
so I have updated that code as well as removed the Labels field from
the struct itself. The temporary hard-coded values for the labels
has also been updated for readability and to show how it would be
properly represented in storage.

Additionally, on the base.html file I have un-commented the
references to bootstrap-select (which were likely creating issues
with the display) and changed the columns back to 2-8-2 spacing.

Note: WIP check-in
Note: Progress on the frontend website content