Attempt to update/fix console rendering
Description:
It was not working on my machine but I was trying to get the
dropdowns to populate correctly. However, I have added in the
section title as well as moved the "selected" logic into the
correct spot for each dropdown option.

I've also changed the hard-coded test values in handler.go to
reflect the correct selection arrangement (since only one
selection is allowed for the "type" dropdowns).

Note: This code was not working on my machine