Github API Calls Working + CSRF
Description: The Github API Calls are working and in a final state. The CSRF is working but the token is underutilized I believe we are still exposed between repos and console. Ideally the repo ids received  in the console should be verified against the CSRF token. In the repos handler the token can be mapped to a set of repo ids and stored in the session. The session can be retreived in the console handler via the incoming CSRF token. This would even eliminate the need to actually POST the repo ids. I am also including console2.html intentionally as a working example of a concept. Delete if you like.