Implement template execution error handling
Description:
In order to ensure a full write out to the response writer, buffers
have now been included on several of the handlers (I didn't do this
for the console handlers since there are several being used for
testing purposes). There are also several minor cosmetic/formatting
updates included in this check-in.

Note: WIP check-in
Note: Unit tests are passing but incomplete