Add HTML string check for completeness
Description:
This is to help ensure that the templates are executing properly and
presenting the appropriate HTML. This is just the first addition (as
the other handlers/scenarios will need to be added) towards getting
more coverage in that area.

There might be a way to read the complete and base HTML files into
memory, parse them a bit, and then compare them against the template
output (but this might be more work than is worth it).

Note: WIP check-in
Note: Unit tests are passing but incomplete
Note: Additional unit tests (similar to this) have yet to be written
      for their respective handlers