Add another HTML unit test
Description:
This is an additional unit test for the frontend HTML template
execution; this is "outside" of the handler unit tests.

Refactoring the actual handlers so that the actual template execute
method output can be captured and independently unit tested should
be addressed in the near future.

Note: Unit test update
Note: Unit tests are passing but not comprehensive
Note: This includes some NOTEs/TODOs in the code