Update frontend server
Description:
Some of the methods and functions have been cleaned up to allow the
PROD flags to guard variables and avoid potential errors.

Additionally, the redirect handler is now a closure to allow for
easier access to setting the URL in integration/unit tests.

Note: WIP on #51
Note: golint and formatting requirements have been met
Note: Unit tests have NOT been implemented on server.go
Note: In testing on local the local machine the redirect and
      "direct" landing pages rendered appropriately