Update website assets for login user flow
Description:
Several files needed to have the href's updated (to direct around
the website).

Additionally, I've added a "submit" button to the repos page
rather than rely on a JS update feature (note that the first
dropdown is auto-selected on the page).

Lastly, I've pulled the CSS for the but landing page SIGNUP button
because it doesn't seem to be rendering correctly/the right color.
I'll circle back around to this at some point.

Note: Known error at the "submit" on repos.html (gob related)
Note: Pages appear to render/flow correctly in Firefox
Note: Some of this still need to be updated (terms/privacy links)