Reimplementing buffer write for template execution
Description:
I'm still pretty darn sure we cannot write out directly to the
response writer twice (which I think can happen in the case an error
is encountered); in the current structure, we now avoid doing that
by temporarily storing the execution output in a buffer and then
checking for the existence of an error.

This feature has been implemented/removed recently but I want to
make sure it's not necessary before removing it again.

Note: WIP check-in
Note: Progress on #51
Note: Unit tests are passing but not comprehensive