Update backend to comply with int64 repo ID change
Description:
Since the go-github library has transitioned to using int64 for repo
IDs, it was necessary to propagate that change throughout the
backend logic. This mostly applies to map references but also to
some function/method arguments/return values.

I've also made some updates to golint requirements (many more still
need to be made but that can be handled later).

Note: WIP check-in
Note: Progress on #51
Note: Unit tests are passing (as far as I can tell)
Note: There is a known failure ("r declared and not used") which is
      part of some in-progress code changes