Change template execution method
Description:
I don't think the ExecuteTemplate method is the correct one in the
current design of the handler. In this case, Execute should be fine
since we are not creating (at least not materially in this case)
different templates that need to be differentiated (plus I don't
think "base.html" is the correct way to reference the template).

I've made the change and it passes unit tests; in the future, I
would like to explore the possibility of refactoring out these
methods because I think we can generate all of the necessary
templates once and then call them by their names (in which case
we would switch back to ExecuteTemplate).

Note: WIP check-in
Note: Minor update (no apparent material impact)
Note: Unit tests are passsing
Note: Progress on #51