Incomplete label reading logic
Description:
This is the logic required for pulling the selected labels out of
the ingestor to populate into the labelmaker model.

I chose to put the labels as a field on the ArchRepo since we are
currently storing the model there as well (this will change at a
future point). The logic in the activateHandler is also querying
the database but I'm not certain that is the best way to get the
data needed.

Note: WIP check-in
Note: This code has not been tested and is likely incomplete
Note: There are known panics in unit tests included in this commit