Adjust label reading code and query
Description:
Some of the stuff I had in an earlier check-in needed to be rebuilt
(as expected) and this is most of that.

Additional fields needed to be included on the configuration struct
as well as logic for population/reading those variables. The SQL
query has also been written (essentially a copy-paste from the
existing options). The end result of this is that the ReadLabels
method has been depreciated and its logic is now included in the
ReadHeuprConfigSettings method.

Note: WIP check-in
Note: Progress on #51
Note: This code has not been unit tested
Note: This includes a known panic from another portion of the code
      unit tests