Minor unit test for dispatcher
Description:
I'm not entirely sure how best to test this particular method since
there is no returned value/changes that are realized outside of the
function scope (since the goroutines for the workers are spun up
inside the method itself). However, this does appear to be passing
and I may circle back to this one to figure out a better and more
comprehensive unit test.

Note: Unit test update
Note: This check-in includes a known unit test panic