Update model initialization method unit test
Description:
In the production code, I've moved the NewClient method for the
Google language API into it's own function (as a variable for
unit testing if needed).

I've also removed the error value returned on the NewModel method.

Note: Unit test check-in
Note: Panic (known) is still encountered in this directory tests