Add Travis-CI config file
Description:
This is a very simple config that will definitely need to be
adjusted going forward. Right now, I won't set it so that if it
breaks it will fail builds but I'll make that change as soon as
we get the repo/Travis to a fully passing/functional state.